requirements.txt
discord.py==2.7.0
aiosqlite
python-dotenv
datetime
colorama